resolv: Fix tests by aligning hand crafted queries
When testing OpenRISC I get a bus error in res_send.  This is due to the
buf being cast to a (HEADER *) and trying the res_send code trying to read
different bits of the HEADER struct including 16-bit id etc.

On OpenRISC reads of 16-bits and 32-bits from structures need to be 2
byte and 4 byte aligned, respectively.

To fix this we can align the hand crafted queries.