open64: Force O_LARGEFILE on 32-bit architectures
When running tests on OpenRISC which has 32-bit wordsize but 64-bit
timesize it was found that O_LARGEFILE is not being set when calling
open64.  For 64-bit architectures the O_LARGEFILE flag is generally
implied by the kernel according to force_o_largefile.  However, for
32-bit architectures this is not done.

For this patch I add an extra condition to the EXTRA_OPEN_FLAGS ifdef
logic to allow setting O_LARGEFILE on 32-bit architectures.

Tested on the OpenRISC the build works and timezone/tst-tzset passes
which was failing before.  I would expect this also would fix arc.