or1k: Support using a different LD for hard-float
OpenRISC ABI is not actually different for hard-float (all function
calling conventions and types are the same).  The only thing hard-float
does is enable hard-float instruction generation which could be emulated
via illegal instruction hooks.

So maybe this patch it not needed at all.  Leaving it on this branch
in case we need to bring it back.